Fix 32-bit overflow in v2 BCS length checks and related tests
Compare slice lengths via uint64 against maxU32Length instead of
comparing int len() to 0xFFFFFFFF. Use the same constant in Uleb128
overflow checks for consistency.

Also fix v1 BDD and conversion tests that used int literals too large
for 32-bit platforms.

+++ typeConversion_test.go
@@ -1296,13 +1296,13 @@ func TestConvertToI64(t *testing.T) {
 	}{
 		{
 			name:    "int positive",
-			input:   100000000000,
+			input:   int64(100000000000),
 			want:    100000000000,
 			wantErr: false,
 		},
 		{
 			name:    "int negative",
-			input:   -100000000000,
+			input:   int64(-100000000000),
 			want:    -100000000000,
 			wantErr: false,
 		},
